Clean up specialization constants in OpenGL scene renderer
Also add support for disabling omni and spot lights when not used

+++ drivers/gles3/rasterizer_scene_gles3.h
@@ -392,10 +392,10 @@ private:
 		GeometryInstanceSurface **elements = nullptr;
 		int element_count = 0;
 		bool reverse_cull = false;
-		uint32_t spec_constant_base_flags = 0;
+		uint64_t spec_constant_base_flags = 0;
 		bool force_wireframe = false;
 
-		RenderListParameters(GeometryInstanceSurface **p_elements, int p_element_count, bool p_reverse_cull, uint32_t p_spec_constant_base_flags, bool p_force_wireframe = false) {
+		RenderListParameters(GeometryInstanceSurface **p_elements, int p_element_count, bool p_reverse_cull, uint64_t p_spec_constant_base_flags, bool p_force_wireframe = false) {
 			elements = p_elements;
 			element_count = p_element_count;
 			reverse_cull = p_reverse_cull;
